backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
pydantic==2.9.2
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
bcrypt==4.2.0
cryptography==43.0.1
paramiko==3.4.1
python-multipart==0.0.12
pywinrm==0.4.3
requests-ntlm==1.3.0
impacket==0.13.0
apscheduler==3.10.4
pymysql==1.1.1
psycopg2-binary==2.9.10
redis==5.2.1
pymongo==4.10.1
pyodbc==5.0.1
structlog==24.4.0
slowapi==0.1.9
alembic==1.13.3
pysnmp==4.4.12
pytest==8.3.4
pytest-asyncio==0.25.2
python-dateutil==2.9.0
openpyxl==3.1.5
python-dotenv==1.0.1
